% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apache.R
\name{apache}
\alias{apache}
\title{This is function calculates the apache II score}
\usage{
apache(conn, inData, outData)
}
\arguments{
\item{conn}{the connection from connect_db}

\item{inData}{the table name of the input containing visit_no and index_date}

\item{outData}{the table name for the output}
}
\description{
This is function calculates the apache II score
}
\examples{
\dontrun{
apache(conn, "user.cohort", "user.output")
}
}

