% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charlson.R
\name{charlson}
\alias{charlson}
\title{Charlson score from visit_no}
\usage{
charlson(conn, pt_list, out_file = NA, look_back = 5)
}
\arguments{
\item{conn}{the connection from connect_db}

\item{look_back}{the time looking back 1 or 5 years}

\item{visit_no}{the visit_no from cds.cds_visit}
}
\description{
This function calculates the Charlson score based on visit_no
}
\examples{
\dontrun{
charlson(conn, "pt list.csv")
}
}

