% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_sql.R
\name{commit_sql}
\alias{commit_sql}
\title{commit .sql file to database}
\usage{
commit_sql(conn, sql_file, file_flag = T)
}
\arguments{
\item{conn}{the connection from connect_db}

\item{sql_file}{the sql file without any comments}

\item{file_flag}{default TRUE, if FALSE, the sql file itself is treated as a string}
}
\description{
commit .sql file to database
}
\examples{
\dontrun{
commit_sql(conn, "Z:/EMR/bearhunt.sql")
commit_sql(conn, "select * from session.ca", file_flag = F)
}

}

