% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\title{Connect to db2}
\usage{
connect_db(driver, host, port, dbname, user, pw)
}
\arguments{
\item{driver}{the path to the jdbc driver}

\item{host}{The host name of the database.}

\item{port}{the port number of the database}

\item{dbname}{the name of the database}

\item{user}{the user name to log in}

\item{pw}{the password for the user}
}
\description{
This function allows to connect to db2 database
}
\examples{
connect_db("C:/db2jcc4.jar", 'xx.xxx.org' port = "12345", user = "user", pw = "pw")
}

