% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emr4wash-package.R
\docType{package}
\name{emr4washu}
\alias{emr4washu}
\alias{emr4washu-package}
\title{emr4washu: A package for working with eletrical medical records(EMR).}
\description{
This package provide easier UI for mining the EMR including:
lab, charlson, apache and commit sql.
}
\seealso{
\url{https://github.com/dajuntian/emr4washu/blob/master/README.md#emr4washu}
for examples
}

