% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_cds.R
\name{lab_cds}
\alias{lab_cds}
\title{This function fetches labs based on visit_no from cds.labs}
\usage{
lab_cds(conn, inData, outData)
}
\arguments{
\item{conn}{the connection returned from connect_db}

\item{inData}{the table name of the input containing visit_no}

\item{outData}{the table name for the output}
}
\description{
This function fetches labs based on visit_no from cds.labs
}
\examples{
\dontrun{
lab(conn, "user.cohort", "user.output")
}
}

