% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charlson.R
\name{charlson}
\alias{charlson}
\title{Charlson score from visit_no}
\usage{
charlson(conn, pt_list)
}
\arguments{
\item{visit_no}{the visit_no from cds.cds_visit}
}
\description{
This function calculates the Charlson score based on visit_no
}
\examples{
\dontrun{
charlson(conn, "pt list.csv")
}
}

