% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charlson.R
\name{charlson}
\alias{charlson}
\title{Charlson score from visit_no}
\usage{
charlson(conn, inData, outData)
}
\arguments{
\item{conn}{the connection from connect_db}

\item{inData}{the table name for input, must have visit_no}

\item{outData}{the table name for the output}
}
\description{
This function calculates the Charlson score based on visit_no
}
\examples{
\dontrun{
charlson(conn, "user.incohort", "usr.out_cohort")
}
}

